feat(wft): wire the wetland tile through mksrf, restart, BGC and LAI
Ports the proven wiring of the first WFT branch onto the lean base and
adapts it to the eight-class design:

- BGC crop-test bounds (ae46bca8 subset): 47 'is this a crop' sites gain
  the npcropmax upper bound so WFT indices 79..86 are never treated as
  CFTs. Behaviour-neutral until wetland tiles exist.
- Ownership routing (db6acdd0 subset) and restart gating (843ec837):
  patch_has_pft replaces the scattered patchtype==0 tests, Thermal's
  and/or precedence fix rides along, PFT-level restart reads and writes
  are gated for tile-less patches.
- patchtypes becomes integer (04c3026c, 514f8b69, 8903a00c), nint()
  wrappers dropped; the single-point PCT guard lands in its final form.
- Tile construction (ae46bca8), site class channel (61bc88ca) and the
  per-PFT LAI chain (2a1d47b1): one WFT tile per wetland patch, typed
  npwetlandmin+class-1 from SITE_wetland_class (validated 1..8, printed
  at startup); gridded runs refuse loudly until the class-map
  aggregation is wired. The retired seven-code axis shared the variable
  name with different meanings; comments and attributes now carry the
  eight-class legend.
- Soil ownership is hard-coded, not switched: the dormant WETLAND_PFT
  return in the decomposition shim is removed rather than renamed --
  activating it would have handed wetland soil to bgc_driver, the
  collapse the first branch hit at 70 of 77 sites. No wetland_bgc_soil
  switch exists until a BGC-on experiment proposes one.

Macro spelling updated to LULC_IGBP_WFT throughout the ported hunks.
Full gnu build verified (surface data, initial data, colm, postprocess).
SAFETY: binaries from this commit must not run against site srfdata
made before the tile existed -- regenerate srfdata first.

+++ mksrfdata/Aggregation_MethanePH.F90
@@ -95,8 +95,8 @@ SUBROUTINE Aggregation_MethanePH (dir_rawdata, dir_model_landdata, lc_year)
          DO ipatch = 1, numpatch
             IF (landpatch%settyp(ipatch) >= 1 .and. &
                 landpatch%settyp(ipatch) <= size(patchtypes)) THEN
-               requires_spatial_ph = nint(patchtypes(landpatch%settyp(ipatch))) == 0 .or. &
-                  nint(patchtypes(landpatch%settyp(ipatch))) == 2
+               requires_spatial_ph = patchtypes(landpatch%settyp(ipatch)) == 0 .or. &
+                  patchtypes(landpatch%settyp(ipatch)) == 2
                IF (requires_spatial_ph) relevant_local = relevant_local + 1
             ENDIF
          ENDDO
@@ -253,8 +253,8 @@ SUBROUTINE Aggregation_MethanePH (dir_rawdata, dir_model_landdata, lc_year)
             requires_spatial_ph = .false.
             IF (landpatch%settyp(ipatch) >= 1 .and. &
                 landpatch%settyp(ipatch) <= size(patchtypes)) THEN
-               requires_spatial_ph = nint(patchtypes(landpatch%settyp(ipatch))) == 0 .or. &
-                  nint(patchtypes(landpatch%settyp(ipatch))) == 2
+               requires_spatial_ph = patchtypes(landpatch%settyp(ipatch)) == 0 .or. &
+                  patchtypes(landpatch%settyp(ipatch)) == 2
             ENDIF
             IF (requires_spatial_ph .and. .not. methane_ph_valid(ipatch)) &
                invalid_local = invalid_local + 1
